removed dependence on app service
the way the code had been written prior, it was specifically looking for app service-injected environment variables that won`t be there if a customer is deploying the app to Azure Container Apps. In that case, one could use configuration settings if one desired, but for now, the two ports - silo and gateway - are presumed to be the defaults. This way it should work on either app service or container apps without modification.